Renamed kernel_requirements.go to errors.go and extracted the sysOpenAt2Supported function
