--- pkg/network/usm/kernel_requirements.go
@@ -1,40 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-//go:build linux_bpf
-// +build linux_bpf
-
-package usm
-
-import (
-	ddebpf "github.com/DataDog/datadog-agent/pkg/ebpf"
-	"github.com/DataDog/datadog-agent/pkg/network/config"
-	"github.com/DataDog/datadog-agent/pkg/util/kernel"
-	"github.com/DataDog/datadog-agent/pkg/util/log"
-)
-
-// errNotSupported indicates that the current host doesn't fullfil the
-// requirements for USM monitoring
-type errNotSupported struct {
-	error
-}
-
-func (e *errNotSupported) Unwrap() error {
-	return e.error
-}
-
-func sysOpenAt2Supported(c *config.Config) bool {
-	missing, err := ddebpf.VerifyKernelFuncs(doSysOpenAt2.section)
-	if err == nil && len(missing) == 0 {
-		return true
-	}
-	kversion, err := kernel.HostVersion()
-	if err != nil {
-		log.Error("could not determine the current kernel version. fallback to do_sys_open")
-		return false
-	}
-
-	return kversion >= kernel.VersionCode(5, 6, 0)
-}
